Allow overriding some parts
Instead of using all parts of the library to run the OPC server there are scenarios where
people only wants to use parts of the library.

For example I'm writing an OPC proxy, I can't assume before starting the server how many
channels there will be, so I need to be able to have my own dispatcher.

Also I fixed the comment style for functions to match the go style.

Made ReadOPC a standalone function and exported it so that can be used standalone.